content-summarizer/requirements.txt
# Content Summarizer 依赖

# 核心依赖
yt-dlp>=2024.1.0
PyYAML>=6.0

# AI 调用
openai>=1.0.0
aiohttp>=3.9.0

# 可选：飞书 SDK（后续开发）
# feishu-sdk>=0.0.1
